requirements.txt
langchain>=0.2.0
langchain-community>=0.0.19       
langchain-google-genai>=0.0.7
google-generativeai>=0.3.2

# Vector store y procesamiento de PDFs
faiss-cpu
PyPDF2>=3.0.1
pymupdf
unstructured
chromadb

# Utilidades adicionales
wikipedia
tqdm
python-dotenv
pyyaml
pandas
numpy
